+++ packages/balm-core/index.d.ts
@@ -57,11 +57,7 @@ export interface BalmLoaders {
 
 export interface PostcssLoaderOptions {
   exec?: boolean;
-  parser?: string | object;
-  syntax?: string | object;
-  stringifier?: string | object;
-  config?: object;
-  plugins: object[] | Function; // NOTE: The same to `styles.postcssPlugins`
+  postcssOptions?: object | Function; // NOTE: `postcssOptions.plugins` is the same to `styles.postcssPlugins`
   sourceMap: string | boolean;
 }